Move POS product files to correct package
